part_2 complete
Implemented the invite function. When user click on the link send by
the invitation email, he will be directed to a registration page
(implemented under act = ‘register’). A user with corresponding
permission can now give opinion data. The geolocation part of the data
of the opinion data is automatically filled by the browser.